moved json processing to utils
