+++ apps/gitness/src/pages/repo-sandbox/repo-sandbox-branch-rules-container.tsx
@@ -1,22 +1,23 @@
-import { RepoBranchSettingsRulesPage, RepoBranchSettingsFormFields, Rule, BypassUsersList } from '@harnessio/playground'
+import { RepoBranchSettingsRulesPage, RepoBranchSettingsFormFields, BypassUsersList } from '@harnessio/playground'
 import { useGetRepoRef } from '../../framework/hooks/useGetRepoPath'
 import { useState } from 'react'
 import {
   useRuleAddMutation,
   useListPrincipalsQuery,
   useListStatusCheckRecentQuery,
-  EnumMergeMethod,
-  EnumRuleState,
-  RuleAddRequestBody,
+  // EnumMergeMethod,
+  // EnumRuleState,
+  // RuleAddRequestBody,
   useRuleGetQuery,
   RuleGetOkResponse,
   RuleGetErrorResponse,
-  useRuleUpdateMutation,
-  RuleUpdateOkResponse,
-  RuleUpdateErrorResponse
+  useRuleUpdateMutation
+  // RuleUpdateOkResponse,
+  // RuleUpdateErrorResponse
 } from '@harnessio/code-service-client'
 import { useNavigate, useParams } from 'react-router-dom'
 import { useGetSpaceURLParam } from '../../framework/hooks/useGetSpaceParam'
+import { transformDataFromApi, transformFormOutput } from '../../utils/repo-branch-rules-utils'
 
 export const RepoBranchSettingsRulesPageContainer = () => {
   const [preSetRuleData, setPreSetRuleData] = useState()
@@ -24,15 +25,6 @@ export const RepoBranchSettingsRulesPageContainer = () => {
   const repoRef = useGetRepoRef()
   const spaceId = useGetSpaceURLParam()
   const { identifier } = useParams()
-
-  const ruleIds = [
-    'require_latest_commit',
-    'require_no_change_request',
-    'comments',
-    'status_checks',
-    'merge',
-    'delete_branch'
-  ]
 
   if (identifier?.length) {
     useRuleGetQuery(
@@ -49,85 +41,6 @@ export const RepoBranchSettingsRulesPageContainer = () => {
         }
       }
     )
-  }
-  function transformDataFromApi(data: RuleGetOkResponse) {
-    const includedPatterns = data?.pattern?.include || []
-    const excludedPatterns = data?.pattern?.exclude || []
-    const formatPatterns = [
-      ...includedPatterns.map(pat => ({ pattern: pat, option: 'Include' })),
-      ...excludedPatterns.map(pat => ({ pattern: pat, option: 'Exclude' }))
-    ]
-
-    const rules = extractBranchRules(data.definition)
-
-    return {
-      identifier: data.identifier,
-      description: data.description,
-      pattern: data?.pattern?.default,
-      patterns: formatPatterns,
-      rules: rules,
-      state: data.state === 'active',
-      bypass: data?.definition?.bypass?.user_ids,
-      access: '1',
-      default: data?.pattern?.default,
-      repo_owners: data?.definition?.bypass?.repo_owners
-    }
-  }
-
-  const transformFormOutput = (formOutput: RepoBranchSettingsFormFields) => {
-    const rulesMap = formOutput.rules.reduce<Record<string, Rule>>((acc, rule) => {
-      acc[rule.id] = rule
-      return acc
-    }, {})
-
-    const { include, exclude } = formOutput.patterns.reduce<{ include: string[]; exclude: string[] }>(
-      (acc, currentPattern) => {
-        if (currentPattern.option === 'Include') {
-          acc.include.push(currentPattern.pattern)
-        } else if (currentPattern.option === 'Exclude') {
-          acc.exclude.push(currentPattern.pattern)
-        }
-        return acc
-      },
-      { include: [], exclude: [] }
-    )
-
-    const transformed: RuleAddRequestBody = {
-      identifier: formOutput.identifier,
-      type: 'branch',
-      description: formOutput.description,
-      state: (formOutput.state === true ? 'active' : 'disabled') as EnumRuleState,
-      pattern: {
-        default: formOutput.default || false,
-        include,
-        exclude
-      },
-      definition: {
-        bypass: {
-          user_ids: formOutput.bypass,
-          repo_owners: formOutput.repo_owners || false
-        },
-        pullreq: {
-          approvals: {
-            require_code_owners: true,
-            require_latest_commit: rulesMap['require_latest_commit']?.checked || false,
-            require_no_change_request: rulesMap['require_no_change_request']?.checked || false
-          },
-          comments: {
-            require_resolve_all: rulesMap['comments']?.checked || false
-          },
-          merge: {
-            strategies_allowed: (rulesMap['merge']?.submenu || []) as EnumMergeMethod[],
-            delete_branch: rulesMap['delete_branch']?.checked || false
-          },
-          status_checks: {
-            require_identifiers: rulesMap['status_checks']?.selectOptions || []
-          }
-        }
-      }
-    }
-
-    return transformed
   }
 
   const {
@@ -179,12 +92,6 @@ export const RepoBranchSettingsRulesPageContainer = () => {
     }
   )
 
-  // const handleRuleUpdate = (data: RepoBranchSettingsFormFields) => {
-  //   const formattedData = transformFormOutput(data)
-  //   addRule({
-  //     body: formattedData
-  //   })
-  // }
   const handleRuleUpdate = (data: RepoBranchSettingsFormFields) => {
     const formattedData = transformFormOutput(data)
 
@@ -209,53 +116,7 @@ export const RepoBranchSettingsRulesPageContainer = () => {
     addRule: addRuleError?.message || null
   }
 
-  function extractBranchRules(definition: any) {
-    const rules = []
-
-    for (const rule of ruleIds) {
-      let checked = false
-      let submenu: string[] = []
-      let selectOptions: string[] = []
-
-      console.log('cinfirmong ehat strategies allowed looks like?', definition?.pullreq?.merge?.strategies_allowed)
-
-      switch (rule) {
-        case 'require_latest_commit':
-          checked = definition?.pullreq?.approvals?.require_latest_commit || false
-          break
-        case 'require_no_change_request':
-          checked = definition?.pullreq?.approvals?.require_no_change_request || false
-          break
-        case 'comments':
-          checked = definition?.pullreq?.comments?.require_resolve_all || false
-          break
-        case 'status_checks':
-          checked = definition?.pullreq?.status_checks?.require_identifiers?.length > 0
-          selectOptions = definition?.pullreq?.status_checks?.require_identifiers || []
-          break
-        case 'merge':
-          checked = definition?.pullreq?.merge?.strategies_allowed?.length > 0
-          submenu = definition?.pullreq?.merge?.strategies_allowed || []
-          break
-        case 'delete_branch':
-          checked = definition?.pullreq?.merge?.delete_branch || false
-          break
-        default:
-          continue
-      }
-
-      rules.push({
-        id: rule,
-        checked,
-        submenu,
-        selectOptions
-      })
-    }
-
-    return rules
-  }
-
-  console.log('in branch rules general container', preSetRuleData)
+  // console.log('in branch rules general container', preSetRuleData)
 
   return (
     <RepoBranchSettingsRulesPage
